+++ CHANGELOG.md
@@ -4,7 +4,9 @@ Changelog for paco.models
 7.7.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+### Added
+
+- ECS Service has a `capacity_providers` for ECS Capacity Providers.
 
 
 7.7.1 (2020-12-31)